AutoJobAgent/requirements.txt
# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
playwright>=1.42.0

# Document processing
python-docx>=0.8.11
comtypes>=1.3.2
pywin32>=305; platform_system=="Windows"

# Data handling
openpyxl>=3.0.10

# API and dashboard
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.2
websockets>=10.4
pydantic>=1.10.7

# LLM integration
ollama>=0.1.0

# UI and utilities
rich>=13.3.5
watchfiles>=0.19.0

# Testing
pytest>=7.3.1
